Fix out of boundaries access due to typo.

+++ includes/xstd/type_helpers.hpp
@@ -1084,7 +1084,7 @@ namespace xstd
 				}
 				else if constexpr ( L == 1 )
 				{
-					result = T( *( const uint16_t* ) src );
+					result = T( *( const uint8_t* ) src );
 				}
 			}
 			return result;
